+++ test/dummy/db/schema.rb
@@ -11,11 +11,11 @@
 # It's strongly recommended that you check this file into your version control system.
 
 ActiveRecord::Schema[8.1].define(version: 2026_01_09_225213) do
-  create_table "referers", id: :string, force: :cascade do |t|
+  create_table "referers", id: :uuid, force: :cascade do |t|
     t.datetime "created_at", null: false
     t.string "name"
     t.datetime "updated_at", null: false
-    t.string "user_id", null: false
+    t.uuid "user_id", null: false
     t.index ["user_id"], name: "index_referers_on_user_id"
   end